Change how we upload files to avoid race conditions with upload processing
If a number of Jobs are created in a short time frame (or, perhaps, Smartling's
API is just under load), the old approach of uploading a .po file and attaching
it to a Job would fail, with "File locked" when we tried to add a just-uploaded
file to that Job.

A better, more scaleable approach is to switch to the Job Batch API, which
handles processing of uploads asynchronously to a particular Batch, and then
associates them with the Job the Batch belongs to.

This approach works fine, and doesn't stop us from doing other Job operations
(such as setting a Visual Context) while the Batch is processed.